+++ omniisaacgymenvs/scripts/rlgames_demo.py
@@ -31,6 +31,7 @@ import datetime
 import os
 import gym
 import hydra
+import sys
 import torch
 from omegaconf import DictConfig
 import omniisaacgymenvs
@@ -54,6 +55,10 @@ def parse_hydra_configs(cfg: DictConfig):
     time_str = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
 
     headless = cfg.headless
+    # process additional kit arguments and write them to argv
+    if cfg.extras and len(cfg.extras) > 0:
+        sys.argv += cfg.extras
+
     env = VecEnvRLGames(headless=headless, sim_device=cfg.device_id, enable_livestream=cfg.enable_livestream)
 
     # parse experiment directory