+++ omniisaacgymenvs/utils/config_utils/path_utils.py
@@ -82,7 +82,7 @@ def get_experience(headless, enable_livestream, enable_viewport, enable_recordin
             import omniisaacgymenvs
             experience = os.path.abspath(os.path.join(os.path.dirname(omniisaacgymenvs.__file__), '../apps/omni.isaac.sim.python.gym.camera.kit'))
         else:
-            experience = f'{os.environ["EXP_PATH"]}/omni.isaac.sim.python.gym.kit'
+            experience = f'{os.environ["EXP_PATH"]}/omni.isaac.sim.python.kit'
             if headless and not enable_livestream and not enable_recording:
                 experience = f'{os.environ["EXP_PATH"]}/omni.isaac.sim.python.gym.headless.kit'
     else: